+++ CHANGELOG.md
@@ -24,6 +24,18 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://github.com/atlassian-labs/db-replica/compare/release-1.3.0...master
 
+### Added
+- Allow getting cached value from `SuppliedCache` without executing a supplier
+
+### Changed
+
+- Make `PessimisticPropagationConsistency` class final
+- Make `ThrottledCache` class final
+- Make `NoCacheSuppliedCache` class final
+- Add `SuppliedCache#get`
+- Add `NoCacheSuppliedCache#get`
+- Add `ThrottledCache#get`
+
 ## [1.3.0] - 2021-04-21
 [1.3.0]: https://github.com/atlassian-labs/db-replica/compare/release-1.2.6...release-1.3.0